src/requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.7
